by yanniboi: Abstracted minipanel display forms to ctools display forms.
